changed and added routes for the mezcal create, update, and delete

+++ controllers/api/mezcalRoutes.js
@@ -1,21 +1,27 @@
 const router = require("express").Router();
 const Mezcal = require("../../models/Mezcal");
 const withAuth = require("../../utlities/auth");
+const { User } = require("../../models/User");
 
 router.post("/addMezcal", withAuth, async (req, res) => {
   try {
-    const newMezcal = await Mezcal.create({
-      name: req.body.name,
-      style: req.body.style,
-      agave: req.body.agave,
-      abv: req.body.abv,
-      origin: req.body.origin,
-      aroma: req.body.aroma,
-      pallet: req.body.pallet,
-      isManager: req.body.isManager,
-    });
+    if (!req.body.isManager) {
+      return res
+        .status(403)
+        .json({ error: "You do not have permission to add a new Mezcal." });
+    } else {
+      const newMezcal = await Mezcal.create({
+        name: req.body.name,
+        style: req.body.style,
+        agave: req.body.agave,
+        abv: req.body.abv,
+        origin: req.body.origin,
+        aroma: req.body.aroma,
+        pallet: req.body.pallet,
+      });
 
-    res.status(200).json(newMezcal);
+      res.status(200).json(newMezcal);
+    }
   } catch (err) {
     res.status(400).json(err);
   }
@@ -23,20 +29,29 @@ router.post("/addMezcal", withAuth, async (req, res) => {
 
 router.put("/:id", withAuth, async (req, res) => {
   try {
-    const dish = await Mezcal.update(
-      {
-        name: req.body.name,
-        agave: req.body.agave,
-        abv: req.body.abv,
-        isManager: req.body.isManager,
-      }
-      // {
-      //   where: {
-      //     id: req.params.id,
-      //   },
-      // }
-    );
-    res.status(200).json(dish);
+    if (!req.body.isManager) {
+      return res
+        .status(403)
+        .json({ error: "You do not have permission to add a new Mezcal." });
+    } else {
+      const upDateMezcal = await Mezcal.update(
+        {
+          name: req.body.name,
+          style: req.body.style,
+          agave: req.body.agave,
+          abv: req.body.abv,
+          origin: req.body.origin,
+          aroma: req.body.aroma,
+          pallet: req.body.pallet,
+        },
+        {
+          where: {
+            id: req.params.id,
+          },
+        }
+      );
+      res.status(200).json(upDateMezcal);
+    }
   } catch (err) {
     res.status(500).json(err);
   }
@@ -44,18 +59,24 @@ router.put("/:id", withAuth, async (req, res) => {
 
 router.delete("/:id", withAuth, async (req, res) => {
   try {
-    const mezcalData = await Mezcal.destroy({
-      where: {
-        name: req.body.name,
-      },
-    });
+    if (!req.body.isManager) {
+      return res
+        .status(403)
+        .json({ error: "You do not have permission to add a new Mezcal." });
+    } else {
+      const mezcalData = await Mezcal.destroy({
+        where: {
+          name: req.body.name,
+        },
+      });
 
-    if (!mezcalData) {
-      res.status(404).json({ message: "No Mezcal found with this id!" });
-      return;
+      if (!mezcalData) {
+        res.status(404).json({ message: "No Mezcal found with this id!" });
+        return;
+      }
+
+      res.status(200).json(mezcalData);
     }
-
-    res.status(200).json(mezcalData);
   } catch (err) {
     res.status(500).json(err);
   }
